add dictionary item test (and move into subdirectory)
